# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Smart i-n-E Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: currencies
    description: Currency operations
  - name: transactions
    description: Transaction operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /currencies:
    get:
      operationId: getCurrencies
      tags: [currencies]
      summary: Get all currencies with rates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of currencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Currency"

  /currencies/refresh:
    post:
      operationId: refreshCurrencyRates
      tags: [currencies]
      summary: Fetch live exchange rates and update DB
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Rates updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrencyRefreshResponse"
        "503":
          description: Exchange rates API unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions:
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions for the user
      security:
        - bearerAuth: []
      parameters:
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

  /transactions/summary:
    get:
      operationId: getTransactionSummary
      tags: [transactions]
      summary: Get aggregated transaction summary
      security:
        - bearerAuth: []
      parameters:
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
            default: day
      responses:
        "200":
          description: Summary data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionSummary"

  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Soft delete a transaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        mode:
          type: string
          enum: [individual, business]
        home_currency:
          type: string
          default: USD
      required: [name, email, password, mode]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
      required: [token, user]

    UserResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        mode:
          type: string
        home_currency:
          type: string
        created_at:
          type: string
      required: [id, name, email, mode, home_currency, created_at]

    Currency:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        rate_to_usd:
          type: string
        rate_updated_at:
          type: string
      required: [code, name, rate_to_usd, rate_updated_at]

    CurrencyRefreshResponse:
      type: object
      properties:
        updated_count:
          type: integer
        message:
          type: string
      required: [updated_count, message]

    CreateTransactionRequest:
      type: object
      properties:
        account_id:
          type: string
        type:
          type: string
          enum: [income, expense]
        amount_original:
          type: number
        currency_code:
          type: string
        notes:
          type: string
        transacted_at:
          type: string
      required: [account_id, type, amount_original, currency_code]

    Transaction:
      type: object
      properties:
        id:
          type: string
        account_id:
          type: string
        type:
          type: string
        amount_original:
          type: string
        currency_code:
          type: string
        amount_usd:
          type: string
        fx_rate_used:
          type: string
        notes:
          type: string
        transacted_at:
          type: string
        created_at:
          type: string
      required: [id, account_id, type, amount_original, currency_code, transacted_at, created_at]

    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [transactions, total, limit, offset]

    TransactionSummary:
      type: object
      properties:
        total_income_usd:
          type: number
        total_expense_usd:
          type: number
        net_usd:
          type: number
        transaction_count:
          type: integer
        income_rate_per_hour:
          type: number
        expense_rate_per_hour:
          type: number
        profit_status:
          type: string
          enum: [profit, breakeven, loss]
        timeframe:
          type: string
      required: [total_income_usd, total_expense_usd, net_usd, transaction_count, income_rate_per_hour, expense_rate_per_hour, profit_status, timeframe]

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]