# lib/api-zod/src/generated/types/createTransactionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart i-n-E Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateTransactionRequestType } from "./createTransactionRequestType";

export interface CreateTransactionRequest {
  account_id: string;
  type: CreateTransactionRequestType;
  amount_original: number;
  currency_code: string;
  notes?: string;
  transacted_at?: string;
}