# lib/api-zod/src/generated/types/deleteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart i-n-E Tracker API
 * OpenAPI spec version: 0.2.0
 */

export interface DeleteResponse {
  success: boolean;
  message: string;
}