# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart i-n-E Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { RegisterRequestMode } from "./registerRequestMode";

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  mode: RegisterRequestMode;
  home_currency?: string;
}