# lib/api-zod/src/generated/types/transactionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart i-n-E Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { TransactionSummaryProfitStatus } from "./transactionSummaryProfitStatus";

export interface TransactionSummary {
  total_income_usd: number;
  total_expense_usd: number;
  net_usd: number;
  transaction_count: number;
  income_rate_per_hour: number;
  expense_rate_per_hour: number;
  profit_status: TransactionSummaryProfitStatus;
  timeframe: string;
}