# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart i-n-E Tracker API
 * OpenAPI spec version: 0.2.0
 */

export interface UserResponse {
  id: string;
  name: string;
  email: string;
  mode: string;
  home_currency: string;
  created_at: string;
}